+++ ambition_utils/rrule/tests/model_tests.py
@@ -595,7 +595,7 @@ class RRuleTest(TestCase):
 
     def test_generate_dates(self):
         """
-        Test a monthly first day of month rule to catch case of converting tz back using the generate_dates method
+        Assert generate_dates returns the same values as get_dates.
         """
         params = {
             'freq': rrule.MONTHLY,
@@ -610,7 +610,30 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates()
+        
+        self.assertEqual(
+            rule.get_dates(),
+            rule.generate_dates()
+        )
+        
+    def test_get_dates(self):
+        """
+        Test a monthly first day of month rule to catch case of converting tz back using the get_dates method
+        """
+        params = {
+            'freq': rrule.MONTHLY,
+            'interval': 1,
+            'dtstart': datetime.datetime(2017, 1, 1, 22),
+            'bymonthday': -1,
+            'until': datetime.datetime(2017, 5, 1, 22),
+        }
+
+        rule = RRule(
+            rrule_params=params,
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates()
 
         # Check a few dates
         self.assertEqual(len(next_dates), 4)
@@ -623,10 +646,10 @@ class RRuleTest(TestCase):
         rule.pre_save_hooks()
 
         # Get next dates to compare against
-        more_next_dates = rule.generate_dates()
+        more_next_dates = rule.get_dates()
         self.assertEqual(next_dates, more_next_dates)
         
-    def test_generate_dates_with_start_date(self):
+    def test_get_dates_with_start_date(self):
         """
         Test a date generation with a start date.
         """
@@ -642,7 +665,7 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates(num_dates=10, start_date=datetime.datetime(2018, 1, 2))
+        next_dates = rule.get_dates(num_dates=10, start_date=datetime.datetime(2018, 1, 2))
         
         # Check a few dates
         self.assertEqual(len(next_dates), 10)
@@ -651,8 +674,29 @@ class RRuleTest(TestCase):
         self.assertEqual(next_dates[2], datetime.datetime(2018, 4, 1, 2))
         self.assertEqual(next_dates[3], datetime.datetime(2018, 5, 1, 2))
         self.assertEqual(next_dates[4], datetime.datetime(2018, 6, 1, 2))
-
+        
     def test_generate_dates_from_params(self):
+        """
+        Assert generate_dates_from_params returns the same values as get_dates_from_params.
+        """
+        params = {
+            'rrule_params': {
+                'freq': rrule.MONTHLY,
+                'interval': 1,
+                'dtstart': datetime.datetime(2017, 1, 1, 22),
+                'bymonthday': -1,
+                'until': datetime.datetime(2017, 5, 1, 22),
+            },
+            'time_zone': pytz.timezone('US/Eastern'),
+            'num_dates': 3,
+        }
+
+        self.assertEqual(
+            RRule.get_dates_from_params(**params),
+            RRule.generate_dates_from_params(**params)
+        )
+
+    def test_get_dates_from_params(self):
         """
         Tests the class method wrapper
         """
@@ -669,12 +713,12 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates(
+        next_dates = rule.get_dates(
             num_dates=3, 
             start_date=datetime.datetime(2017, 2, 2, 22)
         )
 
-        next_dates_from_params = RRule.generate_dates_from_params(
+        next_dates_from_params = RRule.get_dates_from_params(
             rrule_params=params,
             time_zone=pytz.timezone('US/Eastern'),
             num_dates=3,
@@ -683,9 +727,9 @@ class RRuleTest(TestCase):
 
         self.assertEqual(next_dates, next_dates_from_params)
 
-    def test_model_different_time_zone_end_of_month_generate_dates(self):
+    def test_model_different_time_zone_end_of_month_get_dates(self):
         """
-        Test a monthly first day of month rule to catch case of converting tz back using the generate_dates method
+        Test a monthly first day of month rule to catch case of converting tz back using the get_dates method
         """
         params = {
             'freq': rrule.MONTHLY,
@@ -699,7 +743,7 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates()
+        next_dates = rule.get_dates()
         self.assertEqual(next_dates[0], datetime.datetime(2017, 2, 1, 3))
         self.assertEqual(next_dates[1], datetime.datetime(2017, 3, 1, 3))
         self.assertEqual(next_dates[2], datetime.datetime(2017, 4, 1, 2))
@@ -1020,7 +1064,7 @@ class RRuleTest(TestCase):
         self.assertEqual(rule.rrule_params, clone.rrule_params)
 
         # Assert the generated dates are equal.
-        self.assertEqual(rule.generate_dates(num_dates=4), clone.generate_dates(num_dates=4))
+        self.assertEqual(rule.get_dates(num_dates=4), clone.get_dates(num_dates=4))
 
     @freeze_time('6-15-2022')
     def test_weekly_clone_with_offset(self):
@@ -1054,7 +1098,7 @@ class RRuleTest(TestCase):
 
         # Assert the generated dates are as expected.
         self.assertEqual(
-            rule.generate_dates(num_dates=4),
+            rule.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 22),  # Wednesday
                 datetime.datetime(2022, 6, 24),  # Friday
@@ -1065,7 +1109,7 @@ class RRuleTest(TestCase):
 
         # Two days after each date in the regular series.
         self.assertEqual(
-            future_clone.generate_dates(num_dates=4),
+            future_clone.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 24),  # Friday
                 datetime.datetime(2022, 6, 26),  # Sunday
@@ -1076,7 +1120,7 @@ class RRuleTest(TestCase):
 
         # Two days before each date in the regular series.
         self.assertEqual(
-            past_clone.generate_dates(num_dates=4),
+            past_clone.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 20),  # Monday
                 datetime.datetime(2022, 6, 22),  # Wednesday
@@ -1148,7 +1192,7 @@ class RRuleTest(TestCase):
 
         # Assert the generated dates are as expected for the original.
         self.assertEqual(
-            rule.generate_dates(),
+            rule.get_dates(),
             [
                 datetime.datetime(2022, 10, 15),
                 datetime.datetime(2022, 10, 16),
@@ -1158,7 +1202,7 @@ class RRuleTest(TestCase):
 
         # One day after each date in the regular series.
         self.assertEqual(
-            future_clone.generate_dates(),
+            future_clone.get_dates(),
             [
                 datetime.datetime(2022, 10, 16),
                 datetime.datetime(2022, 10, 17),
@@ -1168,7 +1212,7 @@ class RRuleTest(TestCase):
 
         # One day before each date in the regular series.
         self.assertEqual(
-            past_clone.generate_dates(),
+            past_clone.get_dates(),
             [
                 datetime.datetime(2022, 10, 14),
                 datetime.datetime(2022, 10, 15),