+++ ambition_utils/docs/release_notes.rst
@@ -3,7 +3,8 @@ Release Notes
 
 2.6.0
 -----
-* generate_dates(start_date=) to generate dates after start_date.
+* Added get_dates(start_date=) to calculate dates after start_date. Will replace get_dates.
+* Added get_dates_from_params(start_date=) to calculate dates after start_date from rrule_params. Will replace generate_dates_from_params.
 
 2.5.3
 -----